moved /api to /routes/api because /api is reserved in vercel

+++ server/routes/api/user/updateUser.post.ts
@@ -1,5 +1,5 @@
-import { card } from "~/types/type";
-import { firestore, bucket } from "../../utils/firebase"
+import { user } from "~/types/type";
+import { firestore, bucket } from "../../../utils/firebase"
 import formidable from 'formidable'
 import fs from "fs";
 import stream from 'stream';
@@ -7,30 +7,33 @@ import stream from 'stream';
 
 export default defineEventHandler(async (event) => {
 
+  const uid = getCookie(event, 'uid');
+  if(!uid) return;
+
   const form = formidable({ multiples: true });
 
   const data = await new Promise((resolve, reject) => {
     form.parse(event.node.req, (err, fields, files) => {
-
-      if(files.img){
+      if(files.img) {
         const img = files.img[0]
-        
         resolve({...fields, img});
 
-      } else {
-        resolve({...fields});
       }
-
+      resolve({...fields});
     });
   });
 
-  if(data.img) {
-    
-    const imgPath = data.bucketPath
+  if(data.img){
+
+    const {imgPath} = await $fetch<user>('/api/user/user',
+      { headers: {
+        'Cookie' : `uid=${uid}`
+      } }
+    );
     if(imgPath && imgPath?.length != 0){
       bucket.file(imgPath.toString()).delete();
     }
-  
+
     const img = data.img;
     const originalFilename :string = img.originalFilename;
     const extension = originalFilename.substring(img.originalFilename.lastIndexOf('.') + 1);
@@ -46,44 +49,39 @@ export default defineEventHandler(async (event) => {
       passthroughStream.pipe(file.createWriteStream()).on('finish', () => {
         // The file upload is complete
       });
-  
     }
+  
+    
     streamFileUpload().catch(console.error);
     file.getSignedUrl({
       action: 'read',
       expires: '03-09-2491',
     }).then(signedUrls => {
       // signedUrls[0] contains the file's public URL
-      const card :card = {
-        title: data.title.toString(),
-        desc: data.desc.toString(),
-        question: data.question.toString(),
-        answers: JSON.parse(data.answers.toString()),
+      
+      const newUserData :user = {
+        name: data.username.toString(),
         imgUrl: signedUrls[0],
-        bucketPath: fileName
       }
   
-      const deckId = data.deck.toString()
-      const cardId = data.card.toString()
-      firestore.collection('decks').doc(deckId).collection('cards').doc(cardId).set({
-        ...card
+      // console.log(data.deck.toString())
+      firestore.collection('users').doc(uid).set({
+        ...newUserData,
+        imgPath: fileName
+      }, {
+        merge: true
       })
     });
   } else {
-    // if no img
-    const card :card = {
-      title: data.title.toString(),
-      desc: data.desc.toString(),
-      question: data.question.toString(),
-      answers: JSON.parse(data.answers.toString()),
+
+    const newUserData :user = {
+      name: data.username.toString(),
     }
 
-    const deckId = data.deck.toString()
-    const cardId = data.card.toString()
-    firestore.collection('decks').doc(deckId).collection('cards').doc(cardId).update({
-      ...card
+    firestore.collection('users').doc(uid).set({
+      ...newUserData
+    }, {
+      merge: true
     })
   }
-
-  
 })
